requirements.txt
Django
psycopg[binary]
django-tailwind[reload]
django-htmx
django-ipware
